fixing warning refs #12
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@11047 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/utils/geom/PositionVector.cpp
@@ -316,7 +316,7 @@ PositionVector::getCentroid() const {
 void
 PositionVector::scaleSize(SUMOReal factor) {
     Position centroid = getCentroid();
-    for (int i = 0; i < size(); i++) {
+    for (size_t i = 0; i < size(); i++) {
         myCont[i] = centroid + ((myCont[i] - centroid) * factor);
     }
 }
